+++ packages/desktop-client/src/components/DevelopmentTopBar.tsx
@@ -1,6 +1,6 @@
 import { theme } from '../style';
 
-import { ExternalLink } from './common';
+import ExternalLink from './common/ExternalLink';
 import View from './common/View';
 
 export default function DevelopmentTopBar() {
